pyproject.toml
[tool.poetry]
name = "mockingbird"
version = "0.1.0"
description = "Discord chat bot that selects random players and divides players into random teams."
authors = ["Lucas Connors <lucas@revolutiontech.ca>"]
license = "ISC"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.2"
python-dotenv = "^0.12.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pyupgrade = "^2.1.0"

[tool.black]
line-length = 88
target_version = ["py38"]

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
